Create netdrive_dk.po
Be aware that your readme doesn't state what to do with multiline fields. Ie must we follow line breaks, and shall we leave the same odd "" line 
There is a lot of repetition in the translation. You should make a more clean translate task, you risk that we translate the same item differently, if we see it any major amount of time appart. (And even short ones too).
Finally captialization of letters differs between countries. I have followed your lead, but depending on the use, it will look wrong in some languages.
Final note, english tends to split words, be aware that others tend to put them together. It can be hard to detect what you need especially coupled with the capitalization.